[구조] LifecycleOwnerWrapper 클래스 추가 및 앱 뒤로가기 로직 작성
