+++ common-base/src/main/java/com/mykim/commonbase/BaseFragment.kt
@@ -6,18 +6,14 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.ViewDataBinding
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.lifecycleScope
+import androidx.lifecycle.LifecycleOwner
 import com.bumptech.glide.Glide
 import com.bumptech.glide.RequestManager
 import com.mykim.common_util.ImageViewExtensionWrapper
-import com.mykim.common_util.collect
+import com.mykim.common_util.LifecycleOwnerWrapper
 import com.mykim.common_util.convertDpToPx
-import com.mykim.common_util.onResult
-import com.mykim.core_model.state.ResultUiState
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.flow.Flow
 
-abstract class BaseFragment<Binding: ViewDataBinding> : Fragment(), ImageViewExtensionWrapper {
+abstract class BaseFragment<Binding: ViewDataBinding> : Fragment(), ImageViewExtensionWrapper, LifecycleOwnerWrapper {
 
     private var _binding : Binding? = null
     protected val binding get() = _binding!!
@@ -25,6 +21,8 @@ abstract class BaseFragment<Binding: ViewDataBinding> : Fragment(), ImageViewExt
     protected abstract fun createFragmentBinding(inflater: LayoutInflater, container: ViewGroup?) : Binding
 
     protected open fun initFragment() = Unit
+
+    override fun initLifeCycleOwner(): LifecycleOwner = viewLifecycleOwner
 
     override val requestManager: RequestManager
         get() = Glide.with(this)
